+++ CUDA/src/external/imwarp_flexible/imwarp_flexible.c
@@ -18,14 +18,14 @@
 
 static __inline int64_t mindex3_ml(int64_t i_x, int64_t i_y, int64_t i_z, int64_t n_x, int64_t n_y) { return i_z*n_x*n_y + i_x*n_y + i_y ; }  ;
 
-bool is_affine_3d_matrix(const float A[AFFINE_3D_MATRIX_SIZE])
+bool is_affine_3d_matrix(const double A[AFFINE_3D_MATRIX_SIZE])
 {
 	// A is assumed to be stored col-major, and to be a "row form" affine transform matrix, s.t. y' = x'*A .
 	// Check the last column. It should be [0, 0, 0, 1]'
 	return (A[12] == 0.0f && A[13] == 0.0f && A[14] == 0.0f && A[15] == 1.0f) ;
 }
 
-void print_affine_3d_matrix(const float A[AFFINE_3D_MATRIX_SIZE])
+void print_affine_3d_matrix(const double A[AFFINE_3D_MATRIX_SIZE])
 {
 	// A is assumed to be stored col-major, and to be a "row form" affine transform matrix, s.t. y' = x'*A .
 	int ii;
@@ -37,7 +37,7 @@ void print_affine_3d_matrix(const float A[AFFINE_3D_MATRIX_SIZE])
 #define _m(i,j) (A[i + 4 * j])
 // Macro to get the element of Ainv at row i, col j, assuming Ainv is stored col-major
 #define _minv(i,j) (Ainv[i + 4 * j])
-void invert_affine_3d_matrix(const float  A[AFFINE_3D_MATRIX_SIZE], float Ainv[AFFINE_3D_MATRIX_SIZE])
+void invert_affine_3d_matrix(const double  A[AFFINE_3D_MATRIX_SIZE], double Ainv[AFFINE_3D_MATRIX_SIZE])
 {
 	// A is assumed to be stored col-major, and to be a "row form" affine transform matrix, s.t. y' = x'*A .
 	if (!is_affine_3d_matrix(A))
@@ -67,15 +67,15 @@ void invert_affine_3d_matrix(const float  A[AFFINE_3D_MATRIX_SIZE], float Ainv[A
 
 	double invdet = 1 / det;
 
-	_minv(0, 0) = (float) ((_m(1, 1) * _m(2, 2) - _m(2, 1) * _m(1, 2)) * invdet) ;
-    _minv(0, 1) = (float) ((_m(0, 2) * _m(2, 1) - _m(0, 1) * _m(2, 2)) * invdet) ;
-    _minv(0, 2) = (float) ((_m(0, 1) * _m(1, 2) - _m(0, 2) * _m(1, 1)) * invdet) ;
-    _minv(1, 0) = (float) ((_m(1, 2) * _m(2, 0) - _m(1, 0) * _m(2, 2)) * invdet) ;
-    _minv(1, 1) = (float) ((_m(0, 0) * _m(2, 2) - _m(0, 2) * _m(2, 0)) * invdet) ;
-    _minv(1, 2) = (float) ((_m(1, 0) * _m(0, 2) - _m(0, 0) * _m(1, 2)) * invdet) ;
-    _minv(2, 0) = (float) ((_m(1, 0) * _m(2, 1) - _m(2, 0) * _m(1, 1)) * invdet) ;
-    _minv(2, 1) = (float) ((_m(2, 0) * _m(0, 1) - _m(0, 0) * _m(2, 1)) * invdet) ;
-    _minv(2, 2) = (float) ((_m(0, 0) * _m(1, 1) - _m(1, 0) * _m(0, 1)) * invdet) ;
+	_minv(0, 0) = (double) ((_m(1, 1) * _m(2, 2) - _m(2, 1) * _m(1, 2)) * invdet) ;
+    _minv(0, 1) = (double) ((_m(0, 2) * _m(2, 1) - _m(0, 1) * _m(2, 2)) * invdet) ;
+    _minv(0, 2) = (double) ((_m(0, 1) * _m(1, 2) - _m(0, 2) * _m(1, 1)) * invdet) ;
+    _minv(1, 0) = (double) ((_m(1, 2) * _m(2, 0) - _m(1, 0) * _m(2, 2)) * invdet) ;
+    _minv(1, 1) = (double) ((_m(0, 0) * _m(2, 2) - _m(0, 2) * _m(2, 0)) * invdet) ;
+    _minv(1, 2) = (double) ((_m(1, 0) * _m(0, 2) - _m(0, 0) * _m(1, 2)) * invdet) ;
+    _minv(2, 0) = (double) ((_m(1, 0) * _m(2, 1) - _m(2, 0) * _m(1, 1)) * invdet) ;
+    _minv(2, 1) = (double) ((_m(2, 0) * _m(0, 1) - _m(0, 0) * _m(2, 1)) * invdet) ;
+    _minv(2, 2) = (double) ((_m(0, 0) * _m(1, 1) - _m(1, 0) * _m(0, 1)) * invdet) ;
 
 
 	//translation for bottom row
@@ -123,9 +123,9 @@ int get_number_of_cores()
 voidthread transform_subvolume(void **args) {
 	// Sort out the arguments
 	int64_t* output_dims = (int64_t *)(args[0]);
-	float* output_origin = (float *)(args[1]);
-	float* output_spacing = (float *)(args[2]);
-	float* Ainv = (float*)(args[3]);
+	double* output_origin = (double *)(args[1]);
+	double* output_spacing = (double *)(args[2]);
+	double* Ainv = (double*)(args[3]);
 	float* input_stack = args[4];
 	float* output_stack = args[5];
 	int thread_id = *((int *)(args[6])) ;
@@ -133,36 +133,36 @@ voidthread transform_subvolume(void **args) {
 	bool is_background_black = *((bool *)(args[8]));
 	int n_threads = *((int *)(args[9]));
 	int64_t* input_dims = args[10] ;
-	float* input_origin = args[11] ;
-	float* input_spacing = (float *)(args[12]);
+	double* input_origin = args[11] ;
+	double* input_spacing = (double *)(args[12]);
 
 	// Decode the interpolation_mode
 	//int is_black = (interpolation_mode == 1 || interpolation_mode == 3) ;
 	//int is_cubic = (interpolation_mode == 2 || interpolation_mode == 3) ;
 
-	float acomp0 = Ainv[3];  // the x translation
-	float acomp1 = Ainv[7];  // the y translation 
-	float acomp2 = Ainv[11];  // the z translation
+	double acomp0 = Ainv[3];  // the x translation
+	double acomp1 = Ainv[7];  // the y translation 
+	double acomp2 = Ainv[11];  // the z translation
 	//  Loop through all output image pixel coordinates */
 	for (int64_t i_z_out = thread_id; i_z_out<output_dims[2]; i_z_out = i_z_out + n_threads) {
-		float z_out = output_spacing[2] * (i_z_out + 0.5f) + output_origin[2] ;
-		float bcomp0 = Ainv[2] * z_out + acomp0;
-		float bcomp1 = Ainv[6] * z_out + acomp1;
-		float bcomp2 = Ainv[10] * z_out + acomp2;
+		double z_out = output_spacing[2] * (i_z_out + 0.5f) + output_origin[2] ;
+		double bcomp0 = Ainv[2] * z_out + acomp0;
+		double bcomp1 = Ainv[6] * z_out + acomp1;
+		double bcomp2 = Ainv[10] * z_out + acomp2;
 		for (int64_t i_y_out = 0; i_y_out<output_dims[0]; i_y_out++) {
-			float y_out = output_spacing[1] * (i_y_out + 0.5f) + output_origin[1] ;
-			float ccomp0 = Ainv[1] * y_out + bcomp0;
-			float ccomp1 = Ainv[5] * y_out + bcomp1;
-			float ccomp2 = Ainv[9] * y_out + bcomp2;
+			double y_out = output_spacing[1] * (i_y_out + 0.5f) + output_origin[1] ;
+			double ccomp0 = Ainv[1] * y_out + bcomp0;
+			double ccomp1 = Ainv[5] * y_out + bcomp1;
+			double ccomp2 = Ainv[9] * y_out + bcomp2;
 			for (int64_t i_x_out = 0; i_x_out<output_dims[1]; i_x_out++) {
-				float x_out = output_spacing[0] * (i_x_out + 0.5f) + output_origin[0] ;
-				float x_in = Ainv[0] * x_out + ccomp0;
-				float y_in = Ainv[4] * x_out + ccomp1;
-				float z_in = Ainv[8] * x_out + ccomp2;
+				double x_out = output_spacing[0] * (i_x_out + 0.5f) + output_origin[0] ;
+				double x_in = Ainv[0] * x_out + ccomp0;
+				double y_in = Ainv[4] * x_out + ccomp1;
+				double z_in = Ainv[8] * x_out + ccomp2;
 
-				float i_x_in = (x_in - input_origin[0])/input_spacing[0] - 0.5f ;  // NB: Not necessarily integral.
-				float i_y_in = (y_in - input_origin[1])/input_spacing[1] - 0.5f ;  // NB: Not necessarily integral.
-				float i_z_in = (z_in - input_origin[2])/input_spacing[2] - 0.5f ;  // NB: Not necessarily integral.
+				double i_x_in = (x_in - input_origin[0])/input_spacing[0] - 0.5f ;  // NB: Not necessarily integral.
+				double i_y_in = (y_in - input_origin[1])/input_spacing[1] - 0.5f ;  // NB: Not necessarily integral.
+				double i_z_in = (z_in - input_origin[2])/input_spacing[2] - 0.5f ;  // NB: Not necessarily integral.
 
 				int64_t i_out = mindex3_ml(i_x_out, i_y_out, i_z_out, output_dims[1], output_dims[0]);
 
@@ -176,9 +176,9 @@ voidthread transform_subvolume(void **args) {
 }
 
 //================================================================================================================
-void imwarp_flexible(const float* input_stack, int64_t input_dims[3], float input_origin[3], float input_spacing[3],
-	                 float* output_stack, int64_t output_dims[3], float output_origin[3], float output_spacing[3], 
-					 float A[AFFINE_3D_MATRIX_SIZE], 
+void imwarp_flexible(const float* input_stack, int64_t input_dims[3], double input_origin[3], double input_spacing[3],
+	                 float* output_stack, int64_t output_dims[3], double output_origin[3], double output_spacing[3], 
+					 double A[AFFINE_3D_MATRIX_SIZE], 
 					 bool is_cubic, bool is_background_black)
 {
 	// This assumes imIn is stored Matlab-style in memory: It's a 3D array with dims [n_y n_x n_z], stored col-major.
@@ -210,7 +210,7 @@ void imwarp_flexible(const float* input_stack, int64_t input_dims[3], float inpu
 	// interpolation used when a given output point does not correspond exactly to an on-grid input point.
 
 	// Take the inverse of A once, since we need it at each voxel
-	float Ainv[AFFINE_3D_MATRIX_SIZE];
+	double Ainv[AFFINE_3D_MATRIX_SIZE];
 	invert_affine_3d_matrix(A, Ainv);
 
 	// Determine the number of threads to use
