+++ CUDA/src/external/imwarp_flexible/imwarp_flexible.h
@@ -13,13 +13,13 @@ extern "C"
 
 int get_number_of_cores() ;
 
-bool is_affine_3d_matrix(const float A[AFFINE_3D_MATRIX_SIZE]);
+bool is_affine_3d_matrix(const double A[AFFINE_3D_MATRIX_SIZE]);
 
-void print_affine_3d_matrix(const float A[AFFINE_3D_MATRIX_SIZE]);
+void print_affine_3d_matrix(const double A[AFFINE_3D_MATRIX_SIZE]);
 
-void imwarp_flexible(const float* input_stack, int64_t input_dims[3], float input_origin[3], float input_spacing[3],
-	                 float* output_stack, int64_t output_dims[3], float output_origin[3], float output_spacing[3],
-		             float A[AFFINE_3D_MATRIX_SIZE], 
+void imwarp_flexible(const float* input_stack, int64_t input_dims[3], double input_origin[3], double input_spacing[3],
+	                 float* output_stack, int64_t output_dims[3], double output_origin[3], double output_spacing[3],
+		             double A[AFFINE_3D_MATRIX_SIZE], 
 					 bool is_cubic, bool is_background_black) ;
 
 #ifdef __cplusplus
